requirements.txt
pandas>=2.2.0
openpyxl>=3.1.2
python-docx>=1.1.0
fpdf2>=2.7.8
PyQt6>=6.6.0